+++ packages/live-preview-sdk/src/csm/index.ts
@@ -1,37 +1,7 @@
 import { vercelStegaEncode } from '@vercel/stega';
 import { debug } from '../helpers';
 import jsonPointer from 'json-pointer';
-
-type Source = {
-  field: number;
-  locale: number;
-} & ({ entry: number } | { asset: number });
-
-interface EntitySource {
-  space: number;
-  environment: number;
-  id: string;
-}
-
-type Mappings = Record<string, { source: Source }>;
-
-interface IContentSourceMaps {
-  version: number;
-  spaces: string[];
-  environments: string[];
-  fields: string[];
-  locales: string[];
-  entries: EntitySource[];
-  assets: EntitySource[];
-  mappings: Mappings;
-}
-
-interface GraphQLResponse {
-  data: any;
-  extensions: {
-    contentSourceMaps: IContentSourceMaps;
-  };
-}
+import { EntitySource, GraphQLResponse, Source } from './types';
 
 export class ContentSourceMaps {
   private isUrlOrIsoDate(value: string) {
